+++ test/validatePassword.test.js
@@ -1,17 +1,59 @@
-const { validatePassword } = require('../app/services/userService');
+const { validatePassword } = require('../app/middlewares/userMiddle');
 
 const shortPass = '1234';
 const equalPass = '12345678';
 const longPass = '12345678912346579';
 
 test('No validate password shorter than 8 characters', () => {
-  expect(validatePassword(shortPass)).toBeFalsy();
+  const req = {
+    query: {
+      password: shortPass
+    }
+  };
+  const res = {};
+  const promise = new Promise((resolve, reject) => {
+    validatePassword(req, res, err => {
+      if (!err) {
+        resolve();
+      }
+      reject(err);
+    });
+  });
+  expect(promise).rejects.toThrow();
 });
 
 test('Validate password equal to 8 characters', () => {
-  expect(validatePassword(equalPass)).toBeTruthy();
+  const req = {
+    query: {
+      password: equalPass
+    }
+  };
+  const res = {};
+  const promise = new Promise((resolve, reject) => {
+    validatePassword(req, res, err => {
+      if (!err) {
+        resolve();
+      }
+      reject(err);
+    });
+  });
+  expect(promise).resolves.not.toThrow();
 });
 
 test('Validate password longer than 8 characters', () => {
-  expect(validatePassword(longPass)).toBeTruthy();
+  const req = {
+    query: {
+      password: longPass
+    }
+  };
+  const res = {};
+  const promise = new Promise((resolve, reject) => {
+    validatePassword(req, res, err => {
+      if (!err) {
+        resolve();
+      }
+      reject(err);
+    });
+  });
+  expect(promise).resolves.not.toThrow();
 });
